Update to STA calculation
CombatRally() needed to RefreshUnitAvailableCells() after modifying STA values. Without this if a unit rallied before moving then the new STA values would not be reflected in the cells the unit moved over.

+++ Assets/Scripts/Types/TileType.cs
@@ -9,8 +9,8 @@ public class TileType {
 	// for each public variable here
 
 	public Enums.TileType tileType;
-	public float movementCost;
-    public float staminaCost;
+	public int movementCost;
+    public int staminaCost;
 	public int defenseBonus;
     public string name;
